pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bbot>=2.7.2",
    "dnspython>=2.7.0",
    "google-adk>=1.17.0",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "litellm>=1.79.1",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "pip>=25.3",
    "python-dotenv>=1.0.0",
    "python-docx>=0.8.11",
    "python-whois>=0.9.6",
    "shodan>=1.31.0",
    "termcolor>=3.2.0",
    "asyncpg>=0.31.0",
    "aiohttp>=3.13.2",
    "extractous>=0.3.0",
]
